+++ Engine/Core/LuaScript/LuaBind/Core/LuaBind_Global.cpp
@@ -3,12 +3,22 @@
 // author @zoloypzuo
 #include "Core/LuaScript/LuaScriptManager.h"
 #include "Core/Resource/ResourceManager.h"
+#include "Engine.h"
+#include "GLSLBook/PostEffectPlugin.h"
 
 using namespace Zelo::Core::LuaScript;
 using namespace Zelo::Core::Resource;
 
 void LuaBind_Global(sol::state &luaState) {
-    luaState.set("SCRIPT_DIR", ResourceManager::getSingletonPtr()->getScriptDir().string());
-    luaState.set("RESOURCE_DIR", ResourceManager::getSingletonPtr()->getResourceDir().string());
-    luaState.set_function("print", LuaScriptManager::luaPrint);
+luaState.new_usertype<Plugin>("Plugin",
+"__Dummy", []{}
+);
+luaState.new_usertype<PostEffectPlugin>("PostEffectPlugin",
+sol::base_classes, sol::bases<Plugin>(),
+"__Dummy", []{}
+);
+luaState.set("SCRIPT_DIR", ResourceManager::getSingletonPtr()->getScriptDir().string());
+luaState.set("RESOURCE_DIR", ResourceManager::getSingletonPtr()->getResourceDir().string());
+luaState.set_function("print", LuaScriptManager::luaPrint);
+luaState.set_function("install", [](Plugin *plugin) { Zelo::Engine::getSingletonPtr()->installPlugin(plugin); });
 }
