+++ Engine/Core/RHI/RenderSystem.cpp
@@ -42,4 +42,8 @@ RenderSystem &RenderSystem::getSingleton() {
     return *msSingleton;
 }
 
+void RenderSystem::setRenderPipeline(std::unique_ptr<Core::RHI::RenderPipeline> renderPipeline) {
+    m_renderer = std::move(renderPipeline);
+}
+
