+++ Sandbox/CMakeLists.txt
@@ -1,1 +1,2 @@
+include_directories(./)
 add_subdirectory(GLSLBook)
