+++ Sandbox/GLSLBook/CMakeLists.txt
@@ -1,8 +1,8 @@
 add_library(GLSLBook
-        ForwardRendererPost.cpp
-        ImageProcessing/Bloom.cpp
-        ImageProcessing/Blur.cpp
-        ImageProcessing/Edge.cpp
-        Shadow/ForwardShadowMap.h
-        Shadow/ForwardShadowMapPcf.cpp
-        )
+        PostEffectPipeline.cpp
+#        ImageProcessing/Bloom.cpp
+#        ImageProcessing/Blur.cpp
+#        ImageProcessing/Edge.cpp
+#        Shadow/ForwardShadowMap.h
+#        Shadow/ForwardShadowMapPcf.cpp
+        PostEffectPlugin.cpp PostEffectPlugin.h)
